split_pdf/requirements.txt
PyPDF2==3.0.1
openai>=1.86.0,<2.0.0
python-dotenv==1.0.0
tiktoken>=0.5.0 